Add files via upload
I'm not exactly sure if this is all the way it is supposed to be- I emailed you about it. I've been looking up the code I thought I needed help with to make it look right, but the images are still giving me trouble even after changing them to jpg and the css was being funky as well. I tried to do the website version of css on another website that let me "try it myself" but it still wasn't fully working for me. Thank you!!!